+++ ResultFunctionalXUnit/FunctionalExtensions/Async/OptionTaskExtensionsTest.cs
@@ -6,22 +6,22 @@ using Xunit;
 namespace ResultFunctionalXUnit.FunctionalExtensions.Async
 {
     /// <summary>
-    /// Методы расширения для проверки условий асинхронно. Тесты
+    /// Методы расширения для проверки условий для задачи-объекта. Тесты
     /// </summary>
-    public class WhereAsyncExtensionsTests
+    public class OptionTaskExtensionsTest
     {
         /// <summary>
         /// Условие продолжающее действие. Положительное
         /// </summary>  
         [Fact]
-        public async Task WhereContinueAsync_Ok()
+        public async Task WhereContinueTaskAsync_Ok()
         {
             const string test = "WhereTest";
 
             string testAfterWhere =
-               await test.OptionAsync(testWhere => !String.IsNullOrWhiteSpace(testWhere),
-                testWhere => Task.FromResult(testWhere.ToLowerInvariant()),
-                Task.FromResult);
+               await Task.FromResult(test).OptionTask(testWhere => !String.IsNullOrWhiteSpace(testWhere),
+                testWhere => testWhere.ToLowerInvariant(),
+                testWhere => testWhere);
 
             Assert.Equal(test.ToLowerInvariant(), testAfterWhere);
         }
@@ -30,15 +30,15 @@ namespace ResultFunctionalXUnit.FunctionalExtensions.Async
         /// Условие продолжающее действие. Положительное. Возвращает другой тип
         /// </summary>  
         [Fact]
-        public async Task WhereContinueAsync_Ok_AnotherType()
+        public async Task WhereContinueTaskAsync_Ok_AnotherType()
         {
             const string testParseNumber = "44";
 
             int numberAfterTest =
-               await testParseNumber.OptionAsync(
+               await Task.FromResult(testParseNumber).OptionTask(
                  numberToParse => Int32.TryParse(numberToParse, out _),
-                numberToParse => Task.FromResult(Int32.Parse(numberToParse)),
-                numberToParse => Task.FromResult(0));
+                Int32.Parse,
+                _ => 0);
 
             Assert.Equal(44, numberAfterTest);
         }
@@ -47,14 +47,14 @@ namespace ResultFunctionalXUnit.FunctionalExtensions.Async
         /// Условие продолжающее действие. Негативное
         /// </summary>  
         [Fact]
-        public async Task WhereContinueAsync_Bad()
+        public async Task WhereContinueTaskAsync_Bad()
         {
             const string test = "BadTest";
 
             string testAfterWhere =
-                await test.OptionAsync(testWhere => testWhere.Length == 0,
-                Task.FromResult,
-                testWhere => Task.FromResult(testWhere.ToLower()));
+                await Task.FromResult(test).OptionTask(testWhere => testWhere.Length == 0,
+                testWhere => testWhere,
+                testWhere => testWhere.ToLower());
 
             Assert.Equal(test.ToLowerInvariant(), testAfterWhere);
         }
@@ -63,15 +63,15 @@ namespace ResultFunctionalXUnit.FunctionalExtensions.Async
         /// Условие продолжающее действие. Положительное. Возвращает другой тип
         /// </summary>  
         [Fact]
-        public async Task WhereContinueAsync_Bad_AnotherType()
+        public async Task WhereContinueTaskAsync_Bad_AnotherType()
         {
             const string testParseNumber = "test";
 
             int numberAfterTest =
-                await testParseNumber.OptionAsync(
+                await Task.FromResult(testParseNumber).OptionTask(
                 numberToParse => Int32.TryParse(numberToParse, out _),
-                numberToParse => Task.FromResult(0),
-                numberToParse => Task.FromResult(numberToParse.Length));
+                _ => 0,
+                numberToParse => numberToParse.Length);
 
             Assert.Equal(testParseNumber.Length, numberAfterTest);
         }
@@ -81,13 +81,13 @@ namespace ResultFunctionalXUnit.FunctionalExtensions.Async
         /// Обработка позитивного условия
         /// </summary>  
         [Fact]
-        public async Task WhereOkAsync_LowCaseText()
+        public async Task WhereOkTaskAsync_LowCaseText()
         {
             const string test = "WhereOk";
 
             string testAfterWhere =
-                await test.OptionSomeAsync(testWhere => !String.IsNullOrWhiteSpace(testWhere),
-                testWhere => Task.FromResult(testWhere.ToLowerInvariant()));
+                await Task.FromResult(test).OptionSomeTask(testWhere => !String.IsNullOrWhiteSpace(testWhere),
+                testWhere => testWhere.ToLowerInvariant());
 
             Assert.Equal(test.ToLowerInvariant(), testAfterWhere);
         }
@@ -96,13 +96,13 @@ namespace ResultFunctionalXUnit.FunctionalExtensions.Async
         /// Обработка негативного условия
         /// </summary>
         [Fact]
-        public async Task WhereBadAsync_LowCaseText()
+        public  async Task WhereBadTaskAsync_LowCaseText()
         {
             const string test = "BadTest";
 
             string testAfterWhere =
-                await test.OptionNoneAsync(testWhere => testWhere.Length == 0,
-                    testWhere => Task.FromResult(testWhere.ToLower()));
+                await Task.FromResult(test).OptionNoneTask(testWhere => testWhere.Length == 0,
+                    testWhere => testWhere.ToLower());
 
             Assert.Equal(test.ToLowerInvariant(), testAfterWhere);
         }
